// src/graphql/queries/__generated__/PullRequestFilesQuery.graphql.ts
export type PullRequestFilesQuery$variables = {
  after?: string | null | undefined;
  name: string;
  number: number;
  owner: string;
};
export type PullRequestFilesQuery$data = {
  readonly repository: {
    readonly pullRequest: {
      readonly files: {
        readonly " $fragmentSpreads": FragmentRefs<"PRFilesFragment">;
      } | null | undefined;
    } | null | undefined;
  } | null | undefined;
};
export type PullRequestFilesQuery = {
  response: PullRequestFilesQuery$data;
  variables: PullRequestFilesQuery$variables;
};